pyproject.toml
[project]
name = "gdg-ml-hackweek"
version = "0.1.0"
description = "AI Climbing Coach — multimodal coaching via pose estimation, Gemini, and custom ML models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "mediapipe>=0.10.14",
    "opencv-python>=4.10.0.84",
    "numpy>=1.26.0",
    "google-generativeai>=0.8.0",
    "elevenlabs>=1.9.0",
    "pillow>=10.4.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "aiofiles>=24.1.0",
    "google-genai>=1.65.0",
]
